LayoutTests/http/tests/security/xss-DENIED-regular-propterty-with-iframe-proto-expected.txt


Tests that making other frame window a prototype doesn't expose that window properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS this.__proto__ = targetWindow threw exception TypeError: Cannot set prototype of immutable prototype object.
PASS __proto__ = targetWindow threw exception TypeError: Object.prototype.__proto__ called on null or undefined.
PASS: innerHeight === originalInnerHeight should be 'true' and is.
PASS: this.innerHeight === originalInnerHeight should be 'true' and is.
PASS: successfullyParsed should be 'true' and is.

TEST COMPLETE
